refactored UI (select), routes after rebase
