+++ src/components/Layout/Layout.jsx
@@ -61,3 +61,5 @@ class Layout extends Component {
         )
     }
 }
+
+export default Layout;
