--- src/components/common/dropdown/Dropwdown.scss
@@ -1,34 +1,0 @@
-.dropdown {
-  position: relative;
-  display: inline-block;
-  cursor: pointer;
-}
-
-.dropdown__content {
-  text-align: center;
-  position: absolute;
-  left: -58px;
-  top: 58px;
-  background-color: #f9f9f9;
-  min-width: 100px;
-  box-shadow: 0 8px 16px 0 rgba(0, 0, 0, 0.2);
-  padding: 12px 16px;
-  z-index: 1;
-}
-
-.dropdown__user-icon::after {
-  display: inline-block;
-  width: 0;
-  height: 0;
-  vertical-align: 13px;
-  content: "";
-  border-top: 6px solid;
-  border-right: 6px solid transparent;
-  border-bottom: 0;
-  border-left: 6px solid transparent;
-  transition: transform 0.4s;
-}
-.dropdown__user-icon_flip::after {
-  transform: rotate(180deg);
-  transform-origin: center center;
-}
