+++ src/components/common/input/Input.scss
@@ -15,7 +15,7 @@
   background: 0 0;
   cursor: text;
   outline: transparent;
-  color: #9a9a9a;
+  color: rgba(74,74,74,0.95);
   box-sizing: border-box;
   height: 45px;
   width: 100%;
