+++ src/components/common/logout/Logout.jsx
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 import { Redirect } from "react-router";
-import { Button } from "../button/Button";
+import { Button } from "../index";
 
 export class Logout extends Component {
   state = {
