+++ src/components/forms/ProfileForm/ProfileForm.scss
@@ -1,6 +1,7 @@
 .profile-form {
   width: 50%;
   display: inline-block;
+  color: rgba(74,74,74,0.95);
 }
 
 .form {
