+++ src/pages/SignUpPage/SignUpPage.scss
@@ -92,7 +92,7 @@
   left: 3%;
 }
 
-.input {
+/*.input {
   position: relative;
 
   height: 45px;
@@ -104,7 +104,7 @@
 
   border-bottom: 2px solid #d9d9d9;
   transition: 0.3s ease all;
-}
+}*/
 
 .input:focus {
   border-bottom: 2px solid #7f7f7f;
