+++ src/pages/eventsPage/EventsPage.jsx
@@ -1,10 +1,12 @@
 import React from "react";
-import { Layout } from "../../components/Layout/Layout";
+import Layout from "../../components/Layout/Layout";
 
-export function EventsPage() {
+function EventsPage() {
   return (
     <Layout>
       <h1>Events Page!</h1>
     </Layout>
   );
 }
+
+export default EventsPage;
